Comment and reorder some elements
Created the screen display file to put there the differente components we are going to display with each click.
